# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Demo API
  version: 0.2.0
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /users/{id}:
    put:
      summary: Update user
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Updated
    delete:
      summary: Soft delete user
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Deleted
  /users/{id}/reset-password:
    put:
      summary: Reset a user's password
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Password reset
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...